refactor: split core generator modules
