+++ src/rust/core/src/generator/mod.rs
@@ -1,0 +1,9 @@
+mod modes;
+mod render;
+mod sample;
+mod text;
+
+#[cfg(test)]
+mod tests;
+
+pub use modes::generate_http_files;
